fix(primaryKey): Se cambian primary key de decimal a int por confictor con postgres a la hora de generar las id automaticas. Se crea nueva migracion y se desactiva encriptacion.

+++ src/Bizagi.Microservice.Api.WorldCupPool.Abstractions/Repositories/ILeaderboardRepository.cs
@@ -24,5 +24,5 @@ public interface ILeaderboardRepository
     /// </summary>
     /// <param name="userId">Identificador del usuario cuyo historial se consulta.</param>
     /// <returns>Lista de entradas del historial ordenadas por fecha de partido ascendente.</returns>
-    Task<IEnumerable<UserPredictionHistoryDto>> GetUserHistoryAsync(decimal userId);
+    Task<IEnumerable<UserPredictionHistoryDto>> GetUserHistoryAsync(int userId);
 }
